refactor: Split aibtc-agent-account.test.ts into categorized test files
